requirements.txt
numpy==1.23.5
pandas==1.5.0
torch==1.12.1
torchaudio==0.12.1
tranformers==4.22.2
librosa
praat-parselmouth
boto3